requirement.txt
Django==5.2.2
environs==14.2.0
pillow==11.2.1
django-phonenumber-field==8.1.0
